增加各个模块的按单查询接口

+++ Business/Outbound/ViewModels/SkuBinCodeQty.cs
@@ -4,7 +4,7 @@ namespace dotnet_wms_ef.ViewModels
     {
         public long SkuId { get; set; }
         public string Sku { get; set; }
-        public long BinId { get; set; }
+        public int BinId { get; set; }
         public string BinCode { get; set; }
         public int Qty { get; set; }
     }
